pyproject.toml
[project]
name = "scribeflow"
version = "0.1.0"
description = "Automatic video subtitle generator using Whisper AI and FFmpeg"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faster-whisper>=1.1.1",
    "ffmpeg-python>=0.2.0",
    "pysrt",
    "fastapi",
    "uvicorn[standard]",
    "python-multipart>=0.0.21",
    "httpx",
    "numpy>=1.23.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
]
